Refactor directory to include test package for unit testing.
