+++ src/main/kotlin/com/badger/taskschedulingapp/Main/MyApp.kt
@@ -1,0 +1,13 @@
+package com.badger.taskschedulingapp.Main
+
+import com.badger.taskschedulingapp.Main.staic.Styles
+import com.badger.taskschedulingapp.Main.view.MainView
+import tornadofx.*
+
+class MyApp: App(MainView::class, Styles::class){
+
+
+
+
+}
+
