+++ src/main/kotlin/com/badger/taskschedulingapp/Main/services/PriorityService.kt
@@ -1,0 +1,6 @@
+package com.badger.taskschedulingapp.Main.services
+
+import com.badger.taskschedulingapp.Main.models.Priority
+
+interface PriorityService: CrudService<Priority, Long> {
+}
