+++ src/main/kotlin/com/badger/taskschedulingapp/Main/services/TaskService.kt
@@ -1,0 +1,6 @@
+package com.badger.taskschedulingapp.Main.services
+
+import com.badger.taskschedulingapp.Main.models.Task
+
+interface TaskService: CrudService<Task, Long> {
+}
