+++ src/main/kotlin/com/badger/taskschedulingapp/Main/services/postgresql/PriorityPostgresService.kt
@@ -1,9 +1,9 @@
-package com.badger.taskschedulingapp.services.postgresql
+package com.badger.taskschedulingapp.Main.services.postgresql
 
 
-import com.badger.taskschedulingapp.models.DataSource.Companion.transaction
-import com.badger.taskschedulingapp.models.Priority
-import com.badger.taskschedulingapp.services.PriorityService
+import com.badger.taskschedulingapp.Main.models.DataSource.Companion.transaction
+import com.badger.taskschedulingapp.Main.models.Priority
+import com.badger.taskschedulingapp.Main.services.PriorityService
 
 class PriorityPostgresService: AbstractPostgresqlSerivce<Priority, Long>(), PriorityService{
