+++ src/main/kotlin/com/badger/taskschedulingapp/Main/view/MainView.kt
@@ -1,6 +1,6 @@
-package com.badger.taskschedulingapp.view
+package com.badger.taskschedulingapp.Main.view
 
-import com.badger.taskschedulingapp.staic.Styles
+import com.badger.taskschedulingapp.Main.staic.Styles
 import tornadofx.*
 
 class MainView : View("Hello TornadoFX") {
