--- src/main/kotlin/com/badger/taskschedulingapp/MyApp.kt
@@ -1,13 +1,0 @@
-package com.badger.taskschedulingapp
-
-import com.badger.taskschedulingapp.staic.Styles
-import com.badger.taskschedulingapp.view.MainView
-import tornadofx.*
-
-class MyApp: App(MainView::class, Styles::class){
-
-
-
-
-}
-
