--- src/main/kotlin/com/badger/taskschedulingapp/services/PriorityService.kt
@@ -1,6 +1,0 @@
-package com.badger.taskschedulingapp.services
-
-import com.badger.taskschedulingapp.models.Priority
-
-interface PriorityService: CrudService<Priority, Long> {
-}
