--- src/main/kotlin/com/badger/taskschedulingapp/services/TaskService.kt
@@ -1,6 +1,0 @@
-package com.badger.taskschedulingapp.services
-
-import com.badger.taskschedulingapp.models.Task
-
-interface TaskService: CrudService<Task, Long> {
-}
